Ensure good matchups per gender across teams
Previously, two teams would be considered "even" if one had
3 experienced males and 3 experienced females. Since males
generally mark males and females mark females, this would create
a lot of poor matchups on the field.

With this change, team allocation prefers combinations where
not only is the overall skill the same between teams, but also the
amount of skill per gender.